+++ src/main/java/com/example/play_view/game/GameServiceImpl.java
@@ -8,7 +8,9 @@ import com.example.play_view.genre.GenreRepository;
 import com.example.play_view.publisher.PublisherDTO;
 import com.example.play_view.publisher.PublisherEntity;
 import com.example.play_view.publisher.PublisherRepository;
+import com.example.play_view.validation.EntityNotFound;
 import jakarta.persistence.EntityNotFoundException;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -24,14 +26,15 @@ import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+@Slf4j
 @Service
 public class GameServiceImpl implements GameService{
 
-    private GameRepository gameRepository;
-    private CompanyRepository companyRepository;
-    private PublisherRepository publisherRepository;
-    private GenreRepository genreRepository;
-    private GameDTOMapper gameDTOMapper;
+    private final GameRepository gameRepository;
+    private final CompanyRepository companyRepository;
+    private final PublisherRepository publisherRepository;
+    private final GenreRepository genreRepository;
+    private final GameDTOMapper gameDTOMapper;
 
     @Autowired
     public GameServiceImpl(GameRepository gameRepository, CompanyRepository companyRepository, PublisherRepository publisherRepository, GenreRepository genreRepository, GameDTOMapper gameDTOMapper) {
@@ -44,7 +47,7 @@ public class GameServiceImpl implements GameService{
 
     @Override
     public boolean existById(long id) {
-        return false;
+        return gameRepository.existsById(id);
     }
 
     @Override
@@ -86,11 +89,15 @@ public class GameServiceImpl implements GameService{
 
         Pageable pageable = PageRequest.of(pageNum, pageSize, Sort.by(orderDir, order));
 
-        return gameRepository.findAll(spec, pageable).stream().map(gameDTOMapper).toList();
+        List<GameDTO> gameDTOS = gameRepository.findAll(spec, pageable).stream().map(gameDTOMapper).toList();
+
+        if (gameDTOS.isEmpty()) throw new EntityNotFound("Game not found");
+        return gameDTOS;
     }
 
     @Override
     public List<GameDTO> findById(long id) {
+        if (!existById(id)) throw new EntityNotFound("Game with ID: " + id + " not found");
         return gameRepository.findById(id).stream()
                 .map(gameDTOMapper)
                 .toList();
@@ -100,17 +107,11 @@ public class GameServiceImpl implements GameService{
     public GameDTO saveGame(GameDTO gameDTO) {
         CompanyEntity company = companyRepository.findByCompanyName(gameDTO.company().companyName());
 
-        Set<PublisherEntity> publishers = new HashSet<>(publisherRepository.findAllByPublisherNameIn(
-                gameDTO.publishers().stream()
-                        .map(PublisherDTO::publisherName)
-                        .toList()
-        ));
+        if (company == null) throw new EntityNotFound("Company: " + gameDTO.company().companyName() + " not found");
 
-        Set<GenreEntity> genres = new HashSet<>(genreRepository.findAllByGenreIn(
-                gameDTO.genres().stream()
-                        .map(GenreDTO::genre)
-                        .toList()
-        ));
+        Set<PublisherEntity> publishers = setPublishers(gameDTO);
+
+        Set<GenreEntity> genres = setGenres(gameDTO);
 
         GameEntity game = gameDTOMapper.toEntity(gameDTO);
 
@@ -125,6 +126,42 @@ public class GameServiceImpl implements GameService{
 
     @Override
     public void deleteGameById(long id) {
+        if (!existById(id)) throw new EntityNotFound("Game with ID: " + id + " not found");
         gameRepository.deleteById(id);
     }
+
+    private Set<PublisherEntity> setPublishers(GameDTO gameDTO) {
+        Set<String> publisherNames = gameDTO.publishers().stream()
+                .map(PublisherDTO::publisherName)
+                .collect(Collectors.toSet());
+
+        Set<PublisherEntity> publishers = new HashSet<>(publisherRepository.findAllByPublisherNameIn(publisherNames));
+
+        Set<String> foundPublishersNames = publishers.stream()
+                        .map(PublisherEntity::getPublisherName)
+                        .collect(Collectors.toSet());
+
+        publisherNames.removeAll(foundPublishersNames);
+
+        if (!publisherNames.isEmpty()) throw new EntityNotFound("Publishers not found: " + String.join(", ", publisherNames));
+        return publishers;
+    }
+
+    private Set<GenreEntity> setGenres(GameDTO gameDTO) {
+        Set<String> genreNames = gameDTO.genres().stream()
+                .map(GenreDTO::genre)
+                .collect(Collectors.toSet());
+
+        Set<GenreEntity> genres = new HashSet<>(genreRepository.findAllByGenreIn(genreNames));
+
+        Set<String> foundGenreNames = genres.stream()
+                .map(GenreEntity::getGenre)
+                .collect(Collectors.toSet());
+
+        genreNames.removeAll(foundGenreNames);
+
+        if (!genreNames.isEmpty()) throw new EntityNotFound("Genres not found: " + String.join(", ", genreNames));
+        return genres;
+    }
+
 }
